requirements.txt
flask==2.3.3
Jinja2==3.1.2
Werkzeug==2.3.7
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
mysql-connector-python==8.0.33 